requirements.txt
google-generativeai>=0.3.2
python-dotenv>=1.0.1
transformers>=4.38.0
sentence-transformers>=2.2.2
torch>=2.0.0
faiss-cpu>=1.7.4
pandas>=1.3.5
numpy>=1.21.0
langchain>=0.1.13
langchain-community>=0.0.30
pypdf>=3.0.0
tqdm>=4.64.0
flask
flask_CORS
pyngrok